backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
mysqlclient==2.1.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing-extensions==4.13.2
djangorestframework
djangorestframework-simplejwt
ics
django-cors-headers
